refactor(core)!: deepen module boundaries around pages.json and macro evaluation
- Extract the pages.json read-modify-write into pagesJson.ts behind a single
  writePagesJson entry point: multi-platform #ifdef merging, home reordering
  comment surgery, serialization formatting, file locking and atomic write.
  Platform becomes a parameter instead of the uni-env frozen constant.
- Extract definePage macro evaluation into macro.ts: evaluateDefinePage for
  the scan path and findDefinePageMacro for the transform path, with
  per-script-block failure isolation.
- Add the pipeline seam (createPages/generateAll) with injectable root and
  platform so tests cross the interface instead of poking internal stages.
- Dissolve utils.ts: debug moves to logger.ts, isTargetFile and
  resolvePagesJsonPath move to files.ts, dead writeFileWithLock is removed,
  checkPagesJsonFileSync no longer returns an unused boolean.
- Narrow PageContext: scan/merge stages are private behind scanAndMerge(),
  dead rawOptions field removed, constructor takes an injectable platform.
- Migrate tests to the pipeline seam; pages.json regression tests inject the
  platform through the constructor instead of pre-setting UNI_PLATFORM
  before a dynamic import. Change detection is evaluated lazily inside the
  file lock to keep overlapping in-process updates from writing redundantly.

BREAKING CHANGE: lifecycle hooks receive stage data instead of the whole
PageContext; PageContext scan/merge stages are private; utils exports
(extsToGlob, parseCode, sleep, mergePageMetaDataArray, invalidatePagesModule,
babelGenerate) are removed from the public surface.
